+++ Framework/Events.php
@@ -72,7 +72,7 @@ class EventHandler {
         }
     }
 
-    public function spaceCount($msg) {
+    public function spaceCount() {
         $model = new CoreSpace();
         $nbSpaces = $model->countSpaces();
         $stat = ['name' => 'spaces', 'fields' => ['value' => $nbSpaces]];
@@ -86,7 +86,7 @@ class EventHandler {
         $space = $model->getSpace($msg['space']['id']);
         $statHandler = new Statistics();
         $statHandler->createDB($space['shortname']);
-        $this->spaceCount($msg);
+        $this->spaceCount();
         // create org
         $g = new Grafana();
         $g->createOrg($space['shortname']);
@@ -98,7 +98,7 @@ class EventHandler {
 
     public function spaceDelete($msg) {
         $this->logger->debug('[spaceDelete]', ['space_id' => $msg['space']['id']]);
-        $this->spaceCount($msg);
+        $this->spaceCount();
     }
 
     public function spaceUserCount($msg){