+++ Modules/clients/Model/ClClientUser.php
@@ -42,7 +42,7 @@ class ClClientUser extends Model {
 
     public function delete($id_space ,$id) {
         $sql = "DELETE FROM cl_j_client_user WHERE id=? AND id_space=?";
-        $this->runRequest($sql, array($id));
+        $this->runRequest($sql, array($id, $id_space));
     }