+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 * [module:booking] fix supplementaries deletion
 * [module:invoices] fix set floats for items quantities
 * [module:booking] fix BkCalendarPeriod missing id_space
+* [helpdesk] fix list of tickets and mail origin
+* [configuration] use env var SMTP_FROM instead of MAIL_FROM (deprecated)
+* [module:booking] for mails check user still have a role in space #383
 
 ## 2.1.6